virtio_fs_utils: functions migrated
1. Migrate the related virtio functions
from win_driver_installer_test.py to virtio_fs_utils.py
2. Replace the win_driver_installer_test to
virtio_fs_utils
