+++ provider/virtio_fs_utils.py
@@ -7,6 +7,8 @@ from virttest import data_dir
 from virttest import error_context
 from virttest import utils_misc
 
+from virttest.utils_windows import virtio_win
+
 LOG_JOB = logging.getLogger('avocado.test')
 
 
@@ -231,3 +233,202 @@ def basic_io_test_via_psexec(test, params, vm, usernm, pwd):
             error_context.context("Delete the sub folder on shared directory "
                                   "of guest: ", LOG_JOB.info)
             session.cmd("rmdir /s /q %s" % (fs_dest + "\\" + folder_name))
+
+
+def get_viofs_exe_path(test, params, session):
+    """
+    Get viofs.exe from virtio win iso,such as E:\viofs\2k19\amd64
+
+    :param test: QEMU test object
+    :param params: Dictionary with the test parameters
+    :param session: the session of guest
+    """
+    error_context.context("Get virtiofs exe full path.", test.log.info)
+    media_type = params["virtio_win_media_type"]
+    try:
+        get_drive_letter = getattr(virtio_win, "drive_letter_%s" % media_type)
+        get_product_dirname = getattr(virtio_win,
+                                      "product_dirname_%s" % media_type)
+        get_arch_dirname = getattr(virtio_win, "arch_dirname_%s" % media_type)
+    except AttributeError:
+        test.error("Not supported virtio win media type '%s'", media_type)
+    viowin_ltr = get_drive_letter(session)
+    if not viowin_ltr:
+        test.error("Could not find virtio-win drive in guest")
+    guest_name = get_product_dirname(session)
+    if not guest_name:
+        test.error("Could not get product dirname of the vm")
+    guest_arch = get_arch_dirname(session)
+    if not guest_arch:
+        test.error("Could not get architecture dirname of the vm")
+
+    exe_middle_path = ("{name}\\{arch}" if media_type == "iso"
+                       else "{arch}\\{name}").format(name=guest_name,
+                                                     arch=guest_arch)
+    exe_file_name = "virtiofs.exe"
+    exe_find_cmd = 'dir /b /s %s\\%s | findstr "\\%s\\\\"'
+    exe_find_cmd %= (viowin_ltr, exe_file_name, exe_middle_path)
+    exe_path = session.cmd(exe_find_cmd).strip()
+    test.log.info("Found exe file '%s'", exe_path)
+    return exe_path
+
+
+def create_viofs_service(test, params, session, service="VirtioFsSvc"):
+    """
+    Only for windows guest, to create a virtiofs service
+    :param test: QEMU test object
+    :param params: Dictionary with the test parameters
+    :param session: the session of guest
+    """
+    install_winfsp(test, params, session)
+    exe_path = get_viofs_exe_path(test, params, session)
+    viofs_exe_copy_cmd_default = 'xcopy %s C:\\ /Y'
+    viofs_exe_copy_cmd = params.get("viofs_exe_copy_cmd",
+                                    viofs_exe_copy_cmd_default)
+    if service == "VirtioFsSvc":
+        error_context.context("Create virtiofs own service in"
+                              " Windows guest.",
+                              test.log.info)
+        output = query_viofs_service(test, params, session)
+        if "not exist as an installed service" in output:
+            session.cmd(viofs_exe_copy_cmd % exe_path)
+            viofs_sc_create_cmd_default = 'sc create VirtioFsSvc ' \
+                                          'binpath="c:\\virtiofs.exe" ' \
+                                          'start=auto  ' \
+                                          'depend="WinFsp.Launcher/VirtioFsDrv" ' \
+                                          'DisplayName="Virtio FS Service"'
+            viofs_sc_create_cmd = params.get("viofs_sc_create_cmd",
+                                             viofs_sc_create_cmd_default)
+            sc_create_s, sc_create_o = session.cmd_status_output(
+                viofs_sc_create_cmd)
+            if sc_create_s != 0:
+                test.fail("Failed to create virtiofs service, "
+                          "output is %s" % sc_create_o)
+    if service == "WinFSP.Launcher":
+        error_context.context("Delete virtiofs own service, "
+                              "using WinFsp.Launcher service instead.",
+                              test.log.info)
+        delete_viofs_serivce(test, params, session)
+        session.cmd(viofs_exe_copy_cmd % exe_path)
+        error_context.context("Config WinFsp.Launcher for multifs.",
+                              test.log.info)
+        output = session.cmd_output(params["viofs_sc_create_cmd"])
+        if "completed successfully" not in output.lower():
+            test.fail("MultiFS: Config WinFsp.Launcher failed, "
+                      "the output is %s." % output)
+
+
+def delete_viofs_serivce(test, params, session):
+    """
+    Delete the virtiofs service on Windows guest
+    :param test: QEMU test object
+    :param params: Dictionary with the test parameters
+    :param session: the session of guest
+    """
+    viofs_sc_delete_cmd = params.get("viofs_sc_delete_cmd",
+                                     "sc delete VirtioFsSvc")
+    error_context.context("Deleting the viofs service...", test.log.info)
+    output = query_viofs_service(test, params, session)
+    if "not exist as an installed service" in output:
+        test.log.info("The viofs service was NOT found at the guest."
+                      " Skipping delete...")
+    else:
+        status = session.cmd_status(viofs_sc_delete_cmd)
+        if status == 0:
+            test.log.info("Done to delete the viofs service...")
+        else:
+            test.error("Failed to delete the viofs service...")
+
+
+def start_viofs_service(test, params, session):
+    """
+    Start the virtiofs service on Windows guest
+    :param test: QEMU test object
+    :param params: Dictionary with the test parameters
+    :param session: the session of guest
+    """
+    viofs_sc_start_cmd = params.get("viofs_sc_start_cmd",
+                                    "sc start VirtioFsSvc")
+    error_context.context("Start the viofs service...", test.log.info)
+    test.log.info("Check if virtiofs service is started.")
+    output = query_viofs_service(test, params, session)
+    if "RUNNING" not in output:
+        status = session.cmd_status(viofs_sc_start_cmd)
+        if status == 0:
+            test.log.info("Done to start the viofs service.")
+        else:
+            test.error("Failed to start the viofs service...")
+    else:
+        test.log.info("Virtiofs service is running.")
+
+
+def query_viofs_service(test, params, session):
+    """
+    Query the virtiofs service on Windows guest
+    :param test: QEMU test object
+    :param params: Dictionary with the test parameters
+    :param session: the session of guest
+    :return output: the output of cmd return
+    """
+    viofs_sc_query_cmd = params.get("viofs_sc_query_cmd",
+                                    "sc query VirtioFsSvc")
+    error_context.context("Query the status of viofs service...",
+                          test.log.info)
+    return session.cmd_output(viofs_sc_query_cmd)
+
+
+def run_viofs_service(test, params, session):
+    """
+    Run the virtiofs service on Windows guest
+    :param test: QEMU test object
+    :param params: Dictionary with the test parameters
+    :param session: the session of guest
+    """
+    create_viofs_service(test, params, session)
+    start_viofs_service(test, params, session)
+
+
+def stop_viofs_service(test, params, session):
+    """
+    Stop the virtiofs service on Windows guest
+    :param test: QEMU test object
+    :param params: Dictionary with the test parameters
+    :param session: the session of guest
+    """
+    viofs_sc_stop_cmd = params.get("viofs_sc_stop_cmd", "sc stop VirtioFsSvc")
+    session.cmd(viofs_sc_stop_cmd)
+
+    test.log.info("Query status of the virtiofs service...")
+    output = query_viofs_service(test, params, session)
+    if "RUNNING" in output:
+        test.error("Virtiofs service is still running.")
+    else:
+        test.log.info("Virtiofs service is stopped.")
+
+
+def install_winfsp(test, params, session):
+    """
+    Install the winfsp on Windows guest
+    :param test: QEMU test object
+    :param params: Dictionary with the test parameters
+    :param session: the session of guest
+    """
+    cmd_timeout = params.get_numeric("cmd_timeout", 120)
+    install_path = params["install_winfsp_path"]
+    check_installed_cmd = params.get("check_winfsp_installed_cmd",
+                                     r'dir "%s" |findstr /I winfsp')
+    check_installed_cmd = check_installed_cmd % install_path
+    install_winfsp_cmd = r'msiexec /i WIN_UTILS:\winfsp.msi /qn'
+    install_cmd = params.get("install_winfsp_cmd",
+                             install_winfsp_cmd)
+    # install winfsp tool
+    test.log.info("Install winfsp for windows guest.")
+    installed = session.cmd_status(check_installed_cmd) == 0
+    if installed:
+        test.log.info("Winfsp tool is already installed.")
+    else:
+        install_cmd = utils_misc.set_winutils_letter(session, install_cmd)
+        session.cmd(install_cmd, cmd_timeout)
+        if not utils_misc.wait_for(lambda: not session.cmd_status(
+                check_installed_cmd), 60):
+            test.error("Winfsp tool is not installed.")
